agent-react-tools/requirements.txt

requests>=2.31
pyyaml>=6.0
tqdm>=4.66
rich>=13.7
